+++ src/services/api/scheduling.ts
@@ -3,15 +3,15 @@ import { postRequest } from '../fetch'
 import OneBlinkAppsError from '../errors/oneBlinkAppsError'
 import tenants from '../../tenants'
 import Sentry from '../../Sentry'
-import { getRootElementValue } from '../prepareSubmissionData'
 import { FormSubmissionResult } from '../../types/submissions'
+import { formElementsService } from '@oneblink/sdk-core'
 
 function getBookingQuerystringValue(
   elementId: string | undefined,
   formSubmissionResult: FormSubmissionResult,
 ) {
   if (elementId) {
-    const value = getRootElementValue(
+    const value = formElementsService.getRootElementValueById(
       elementId,
       formSubmissionResult.definition.elements,
       formSubmissionResult.submission,